requirements.txt
python_version >= "3.10"
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
groq==0.4.2
gunicorn==21.2.0
Werkzeug==3.0.1 